pyproject.toml
[tool.poetry]
name = "api-do-zero-etl"
version = "0.1.0"
description = ""
authors = ["Daniel Ribeiro <danizilla@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
tinydb = "^4.8.2"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
logfire = "^2.7.1"
streamlit = "^1.41.1"
opentelemetry-instrumentation-requests = "^0.50b0"
opentelemetry-instrumentation-sqlalchemy = "^0.50b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
